+++ src/annotations/components/annotationForm/AnnotationStreamSelector.scss
