+++ src/annotations/components/annotationForm/AnnotationTimeStopInput.tsx
